# eidos-api/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Eidos Exchange API
  description: |
    Eidos Exchange REST API documentation.

    ## Authentication
    All private endpoints require EIP-712 signature authentication.

    Authorization header format:
    ```
    Authorization: EIP712 {wallet}:{timestamp}:{signature}
    ```

    ## Rate Limiting
    - Per IP: 1000 requests/minute
    - Per Wallet (queries): 100 requests/second
    - Per Wallet (orders): 10 requests/second
    - Per Wallet (withdrawals): 1 request/second

    ## WebSocket
    WebSocket endpoint: `wss://api.eidos.exchange/ws`

    ### Public Channels
    - ticker: Real-time ticker updates
    - depth: Order book depth
    - kline: K-line/candlestick data
    - trades: Recent trades

    ### Private Channels (requires authentication)
    - orders: Order updates
    - balances: Balance updates

  version: 1.0.0
  contact:
    name: Eidos Exchange
    url: https://eidos.exchange
  license:
    name: MIT
servers:
  - url: https://api.eidos.exchange/api/v1
    description: Production server
  - url: https://testnet-api.eidos.exchange/api/v1
    description: Testnet server
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /orders:
    post:
      tags:
        - Orders
      summary: Create order
      description: Creates a new order with signed EIP-712 data
      operationId: createOrder
      security:
        - EIP712Auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      tags:
        - Orders
      summary: List orders
      description: Returns paginated list of orders
      operationId: listOrders
      security:
        - EIP712Auth: []
      parameters:
        - name: market
          in: query
          description: Filter by market
          schema:
            type: string
        - name: side
          in: query
          description: Filter by side
          schema:
            type: string
            enum: [buy, sell]
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [open, partial, filled, cancelled, rejected]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - $ref: '#/components/parameters/StartTimeParam'
        - $ref: '#/components/parameters/EndTimeParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags:
        - Orders
      summary: Batch cancel orders
      description: Cancel multiple orders by IDs or by market/side
      operationId: batchCancelOrders
      security:
        - EIP712Auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchCancelRequest'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchCancelResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/open:
    get:
      tags:
        - Orders
      summary: List open orders
      description: Returns all open (unfilled) orders
      operationId: listOpenOrders
      security:
        - EIP712Auth: []
      parameters:
        - name: market
          in: query
          description: Filter by market
          schema:
            type: string
        - name: side
          in: query
          description: Filter by side
          schema:
            type: string
            enum: [buy, sell]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdersListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /orders/{id}:
    get:
      tags:
        - Orders
      summary: Get order details
      description: Returns details of a specific order
      operationId: getOrder
      security:
        - EIP712Auth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Orders
      summary: Cancel order
      description: Cancels a specific order
      operationId: cancelOrder
      security:
        - EIP712Auth: []
      parameters:
        - $ref: '#/components/parameters/OrderIdParam'
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /deposits:
    get:
      tags:
        - Deposit
      summary: List deposits
      description: Returns paginated list of deposits
      operationId: listDeposits
      security:
        - EIP712Auth: []
      parameters:
        - name: token
          in: query
          description: Filter by token
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, confirmed, credited]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /deposits/{id}:
    get:
      tags:
        - Deposit
      summary: Get deposit details
      description: Returns details of a specific deposit
      operationId: getDeposit
      security:
        - EIP712Auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Deposit ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /withdrawals:
    post:
      tags:
        - Withdrawal
      summary: Create withdrawal
      description: Creates a new withdrawal request
      operationId: createWithdrawal
      security:
        - EIP712Auth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWithdrawalRequest'
      responses:
        '200':
          description: Withdrawal created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      tags:
        - Withdrawal
      summary: List withdrawals
      description: Returns paginated list of withdrawals
      operationId: listWithdrawals
      security:
        - EIP712Auth: []
      parameters:
        - name: token
          in: query
          description: Filter by token
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [pending, processing, completed, failed, cancelled]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /withdrawals/{id}:
    get:
      tags:
        - Withdrawal
      summary: Get withdrawal details
      description: Returns details of a specific withdrawal
      operationId: getWithdrawal
      security:
        - EIP712Auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Withdrawal ID
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalDetailResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Withdrawal
      summary: Cancel withdrawal
      description: Cancels a pending withdrawal
      operationId: cancelWithdrawal
      security:
        - EIP712Auth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Withdrawal ID
          schema:
            type: string
      responses:
        '200':
          description: Withdrawal cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawalDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    EIP712Auth:
      type: apiKey
      in: header
      name: Authorization
      description: |
        EIP-712 signature authentication.
        Format: `EIP712 {wallet}:{timestamp}:{signature}`
  parameters:
...
    OrderIdParam:
      name: id
      in: path
      required: true
      description: Order ID
      schema:
        type: string
    PageParam:
      name: page
      in: query
      description: Page number (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: Items per page (max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    StartTimeParam:
      name: start_time
      in: query
      description: Start time in milliseconds
      schema:
        type: integer
        format: int64
    EndTimeParam:
      name: end_time
      in: query
      description: End time in milliseconds
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 10003
            message: "INVALID_PARAMS"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 10001
            message: "INVALID_SIGNATURE"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 11001
            message: "ORDER_NOT_FOUND"
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Reset timestamp
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RateLimitResponse'
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          description: Response code (0 for success)
        message:
          type: string
          description: Response message
      required:
        - code
        - message
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      example:
        code: 10003
        message: "INVALID_PARAMS"
    RateLimitResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: object
          properties:
            limit:
              type: integer
            window:
              type: string
            retry_after:
              type: integer

    # Pagination
    Pagination:
      type: object
      properties:
        total:
          type: integer
          format: int64
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer

    # Market data schemas
...
    CreateOrderRequest:
      type: object
      required:
        - order_id
        - market
        - side
        - type
        - amount
        - nonce
        - signature
      properties:
        order_id:
          type: string
        market:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [limit, market]
        price:
          type: string
        amount:
          type: string
        time_in_force:
          type: string
          enum: [GTC, IOC, FOK, GTT]
        nonce:
          type: integer
          format: uint64
        signature:
          type: string
          description: EIP-712 signature
        client_order_id:
          type: string
    Order:
      type: object
      properties:
        order_id:
          type: string
        wallet:
          type: string
        market:
          type: string
        side:
          type: string
          enum: [buy, sell]
        type:
          type: string
          enum: [limit, market]
        price:
          type: string
        amount:
          type: string
        filled_amount:
          type: string
        remaining_amount:
          type: string
        avg_price:
          type: string
        status:
          type: string
          enum: [open, partial, filled, cancelled, rejected]
        time_in_force:
          type: string
        nonce:
          type: integer
          format: uint64
        client_order_id:
          type: string
        reject_reason:
          type: string
        created_at:
          type: integer
          format: int64
        updated_at:
          type: integer
          format: int64
    OrderResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          $ref: '#/components/schemas/Order'
    OrderDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          $ref: '#/components/schemas/Order'
    OrdersListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Order'
            pagination:
              $ref: '#/components/schemas/Pagination'
    BatchCancelRequest:
      type: object
      properties:
        order_ids:
          type: array
          items:
            type: string
          maxItems: 100
        market:
          type: string
        side:
          type: string
          enum: [buy, sell]
    BatchCancelResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: object
          properties:
            cancelled:
              type: array
              items:
                type: string
            failed:
              type: array
              items:
                type: string

    # Balance schemas
...
    Deposit:
      type: object
      properties:
        deposit_id:
          type: string
        wallet:
          type: string
        token:
          type: string
        amount:
          type: string
        tx_hash:
          type: string
        block_num:
          type: integer
          format: int64
        status:
          type: string
          enum: [pending, confirmed, credited]
        detected_at:
          type: integer
          format: int64
        confirmed_at:
          type: integer
          format: int64
        credited_at:
          type: integer
          format: int64
        created_at:
          type: integer
          format: int64
    DepositDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          $ref: '#/components/schemas/Deposit'
    DepositsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Deposit'
            pagination:
              $ref: '#/components/schemas/Pagination'

    # Withdrawal schemas
    CreateWithdrawalRequest:
      type: object
      required:
        - token
        - amount
        - to_address
        - nonce
        - signature
      properties:
        token:
          type: string
        amount:
          type: string
        to_address:
          type: string
        nonce:
          type: integer
          format: uint64
        signature:
          type: string
    Withdrawal:
      type: object
      properties:
        withdraw_id:
          type: string
        wallet:
          type: string
        token:
          type: string
        amount:
          type: string
        to_address:
          type: string
        nonce:
          type: integer
          format: uint64
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        tx_hash:
          type: string
        reject_reason:
          type: string
        submitted_at:
          type: integer
          format: int64
        confirmed_at:
          type: integer
          format: int64
        created_at:
          type: integer
          format: int64
    WithdrawalResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          $ref: '#/components/schemas/Withdrawal'
    WithdrawalDetailResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          $ref: '#/components/schemas/Withdrawal'
    WithdrawalsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
      properties:
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Withdrawal'
            pagination:
              $ref: '#/components/schemas/Pagination'

    # Trade schemas
...